fix: add proper attribution for CoreMark benchmark
