+++ crates/benchmark/Cargo.toml
@@ -6,13 +6,6 @@ rust-version = "1.87.0"
 
 [dependencies]
 wasm-interpreter = { path = "../..", default-features = false }
-wasmi = { version = "0.40.0", optional = true }
-wasmtime = { version = "27.0.0", default-features = false, features = [
-  "runtime",
-  "cranelift",
-  "parallel-compilation",
-  "std",
-], optional = true }
 
 # wasm3-rs' PR #46 makes it work with newer rust-bindgen versions again, but it is not yet merged.
 # However, it is already merged in the fork used for the Roc programming language, which could be
@@ -37,7 +30,8 @@ wasmtime = { version = "27.0.0", default-features = false, features = [
 wat = { workspace = true }
 
 [features]
-foreign-interpreters = ["wasmi", "wasmtime"]
+wasmi = []
+wasmtime = []
 
 [[bench]]
 name = "general_purpose"
