--- crates/benchmark/benches/coremark_minimal.rs
@@ -1,8 +1,0 @@
-fn main() {
-    let score_by_runner = benchmark::coremark_minimal::run();
-
-    println!("Runtime: Score (higher is better)");
-    for (runner, score) in score_by_runner {
-        println!("{}: {}", runner, score)
-    }
-}
