+++ crates/benchmark/benches/coremark_minimal/coremark-minimal.wasm
