--- crates/benchmark/src/coremark_minimal/README.md
@@ -1,1 +1,0 @@
-The file `coremark-minimal.wasm` comes from https://github.com/wasm3/wasm-coremark
